refactor(agw): separate dhcp helper cli
This PR separates the DHCP helper CLI into its own folder and creates a separate release for it. The setup is to build a deb artifact with bazel, which is installed alongside Magma. The CLI tool can then be used by mobilityd in the dev VM, deb VM, and the containerized AGW.
