+++ lte/gateway/python/dhcp_helper_cli/tests/BUILD.bazel
@@ -1,0 +1,28 @@
+# Copyright 2022 The Magma Authors.
+
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+load("@python_deps//:requirements.bzl", "requirement")
+load("//bazel:python_test.bzl", "pytest_test")
+
+MAGMA_ROOT = "../../../../../"
+
+LTE_ROOT = "{}lte/gateway/python".format(MAGMA_ROOT)
+
+pytest_test(
+    name = "test_dhcp_helper_cli",
+    size = "small",
+    srcs = ["test_dhcp_helper_cli.py"],
+    imports = [LTE_ROOT],
+    deps = [
+        "//lte/gateway/python/dhcp_helper_cli",
+        requirement("scapy"),
+    ],
+)
