+++ lte/gateway/python/dhcp_helper_cli/tests/test_dhcp_helper_cli.py
@@ -18,14 +18,14 @@ from typing import Optional
 from unittest.mock import patch
 
 import pytest
-from scapy.layers.dhcp import BOOTP, DHCP
-from scapy.layers.l2 import Dot1Q, Ether
-from scripts.dhcp_helper_cli import (
+from dhcp_helper_cli.dhcp_helper_cli import (
     DhcpHelperCli,
     DHCPState,
     MacAddress,
     release_arg_handler,
 )
+from scapy.layers.dhcp import BOOTP, DHCP
+from scapy.layers.l2 import Dot1Q, Ether
 
 MACSTRING = "12:34:56:78:90:AB"
 MAC = MacAddress(MACSTRING)
@@ -48,11 +48,11 @@ DHCP_RELEASE_PKT /= DHCP(options=[("message-type", DHCPState.RELEASE), "end"])
 
 @pytest.fixture()
 def dhcp_helper_cli_fixture(mac: MacAddress = MAC, vlan: int = VLAN, iface=IFACE, ip: Optional[str] = None, server_ip: Optional[str] = None):
-    with patch("scripts.dhcp_helper_cli.AsyncSniffer"):
+    with patch("dhcp_helper_cli.dhcp_helper_cli.AsyncSniffer"):
         return DhcpHelperCli(mac, vlan, iface, ip, server_ip)
 
 
-@patch("scripts.dhcp_helper_cli.sendp")
+@patch("dhcp_helper_cli.dhcp_helper_cli.sendp")
 def test_send_dhcp_discover(sendp_mock, dhcp_helper_cli_fixture):
     dhcp_helper_cli_fixture.send_dhcp_discover()
     pkt = sendp_mock.call_args[0][0]
@@ -64,7 +64,7 @@ def test_send_dhcp_discover(sendp_mock, dhcp_helper_cli_fixture):
     assert dhcp_helper_cli_fixture._state == DHCPState.DISCOVER
 
 
-@patch("scripts.dhcp_helper_cli.sendp")
+@patch("dhcp_helper_cli.dhcp_helper_cli.sendp")
 def test_send_dhcp_request(sendp_mock, dhcp_helper_cli_fixture):
     dhcp_helper_cli_fixture._state = DHCPState.OFFER
     dhcp_helper_cli_fixture.send_dhcp_request()
@@ -77,7 +77,7 @@ def test_send_dhcp_request(sendp_mock, dhcp_helper_cli_fixture):
     assert dhcp_helper_cli_fixture._state == DHCPState.REQUEST
 
 
-@patch("scripts.dhcp_helper_cli.sendp")
+@patch("dhcp_helper_cli.dhcp_helper_cli.sendp")
 def test_send_dhcp_release(sendp_mock, dhcp_helper_cli_fixture):
     dhcp_helper_cli_fixture._state = DHCPState.ACK
     dhcp_helper_cli_fixture.send_dhcp_release()
